+++ components/calendar/examples/multiple-picker-jalali.tsx
@@ -3,19 +3,18 @@
 import * as React from "react";
 import { format } from "date-fns-jalali";
 import { Calendar as CalendarIcon } from "lucide-react";
-import { DateRange } from "react-day-picker";
 
 import { cn } from "@/lib/utils";
 import { Button } from "@/components/ui/button";
-import { CalendarJalali } from "@/components/ui/calendar-jalali";
+import { CalendarJalali } from "../calendar-jalali";
 import {
   Popover,
   PopoverContent,
   PopoverTrigger,
 } from "@/components/ui/popover";
 
-export function RangePickerJalali() {
-  const [dateRange, setDateRange] = React.useState<DateRange>();
+export function MultiplePickerJalali() {
+  const [dateList, setDateList] = React.useState<Date[]>([]);
 
   return (
     <Popover>
@@ -23,26 +22,29 @@ export function RangePickerJalali() {
         <Button
           variant={"outline"}
           className={cn(
-            "w-[280px] justify-start text-left font-normal",
-            !dateRange && "text-muted-foreground"
+            "min-w-[280px] justify-start text-left font-normal",
+            !dateList.length && "text-muted-foreground"
           )}
         >
           <CalendarIcon className="mr-2 h-4 w-4" />
-          {dateRange ? (
-            <span>
-              از {dateRange.from && format(dateRange.from, "PPP")} تا{" "}
-              {dateRange.to && format(dateRange.to, "PPP")}
-            </span>
+          {dateList.length ? (
+            <div>
+              {dateList.map((date, index) => (
+                <span key={index}>{format(date, "PPP")} و </span>
+              ))}
+            </div>
           ) : (
-            <span>{"محدوده مورد نظر را انتخاب کنید"}</span>
+            <span>{"تاریخ ها مورد نظر را انتخاب کنید"}</span>
           )}
         </Button>
       </PopoverTrigger>
       <PopoverContent className="w-auto p-0">
         <CalendarJalali
-          mode="range"
-          selected={dateRange}
-          onSelect={setDateRange}
+          mode="multiple"
+          selected={dateList}
+          onSelect={setDateList}
+          required={true}
+          max={3}
         />
       </PopoverContent>
     </Popover>
