+++ components/calendar/examples/single-picker-jalali.tsx
@@ -6,15 +6,15 @@ import { Calendar as CalendarIcon } from "lucide-react";
 
 import { cn } from "@/lib/utils";
 import { Button } from "@/components/ui/button";
-import { CalendarJalali } from "@/components/ui/calendar-jalali";
+import { CalendarJalali } from "@/components/calendar/calendar-jalali";
 import {
   Popover,
   PopoverContent,
   PopoverTrigger,
 } from "@/components/ui/popover";
 
-export function MultiplePickerJalali() {
-  const [dateList, setDateList] = React.useState<Date[]>([]);
+export function DatePickerJalali() {
+  const [date, setDate] = React.useState<Date>();
 
   return (
     <Popover>
@@ -22,29 +22,24 @@ export function MultiplePickerJalali() {
         <Button
           variant={"outline"}
           className={cn(
-            "min-w-[280px] justify-start text-left font-normal",
-            !dateList.length && "text-muted-foreground"
+            "w-[280px] justify-start text-left font-normal",
+            !date && "text-muted-foreground"
           )}
         >
           <CalendarIcon className="mr-2 h-4 w-4" />
-          {dateList.length ? (
-            <div>
-              {dateList.map((date, index) => (
-                <span key={index}>{format(date, "PPP")} و </span>
-              ))}
-            </div>
+          {date ? (
+            format(date, "PPP")
           ) : (
-            <span>{"تاریخ ها مورد نظر را انتخاب کنید"}</span>
+            <span>{"یک تاریخ را انتخاب کنید"}</span>
           )}
         </Button>
       </PopoverTrigger>
       <PopoverContent className="w-auto p-0">
         <CalendarJalali
-          mode="multiple"
-          selected={dateList}
-          onSelect={setDateList}
-          required={true}
-          max={3}
+          mode="single"
+          selected={date}
+          onSelect={setDate}
+          autoFocus
         />
       </PopoverContent>
     </Popover>
